+++ simple-stack-example-kotlin/src/main/kotlin/com/zhuinden/simplestackexamplekotlin/NotificationFragment.kt
@@ -1,0 +1,22 @@
+package com.zhuinden.simplestackexamplekotlin
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+import butterknife.ButterKnife
+
+/**
+ * Created by Owner on 2017. 06. 29..
+ */
+
+class NotificationFragment : BaseFragment() {
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
+            inflater.inflate(R.layout.notification_view, container, false)
+
+    override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        ButterKnife.bind(this, view!!)
+    }
+}